Move output writer
